Update to DB and Added AJAX voting script
The CCDB.sql file now automatically drops existing tables and their date
and replaces them with the new variations of the tables. Also, added the
ccVotes table.

The AJAXSendVote.php script can be called to send in a vote on a deck of
cards. it expects deckid, uid, isupv as parameters passed by GET. If
isupv is set to 1 (true) it inserts/updates the vote as an upvote for
the given deck/user combo. If 0, does the same as a downvote.