Clip function script
Added a ccClips table to the DB to store all the decks clipped by each
user.

Added AJAXClipDeck.php to clip a deck. Requires uid and deckid.
Optionally set unClip to cause the script to remove a clip entry, leave
blank otherwise.

Added a function to the deck class that takes in a uid and checks to see
if this deck has been clipped by that user.

Added a function to the user class that gets an array of all the decks
clipped by this user. The array is simply a list of deckid numbers.