Available View Management
Added a new DB table to store deck views by each user.

Added functions to the user class to log a view of a deck into the DB
and get the count of how many views they have earned. Will not add more
than one view per deck/user combo and will not add views for decks that
the user owns. Count is based on the number of decks submitted *
EARN_RATE subtracted from the number of views the user has used.

Adjusted the get display name to only base on alias and not real name.
Added GetReferenceName to return the real name or email address.
Adjusted the login function to create cookies for a user object JSON
string, the user's display name, and the user's reference name.

Also added the GetDeckJSON function back into the deck class.